pyproject.toml
[tool.poetry]
name = "my-cool-app"
version = "0.1.0"
description = ""
authors = ["Karl Dreher"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
jsonschema = "^4.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
